Переделал получение категорий с использованием представлений mysql (Mysql Views). Переделал запросы на запросы с использованием Select метода, предоставляемого библиотекой github.com/jmoiron/sqlx

+++ models/Program.go
@@ -7,5 +7,5 @@ type Program struct {
 	Difficulty int
 	ProgramType string // Тип программы, full-body, split, novice, ...
 	Price int
-	Rating int
+	Rating float32
 }
